Updated axios interceptor and AuthContext
